Player class created
-Made a Player class that tracks the death counter, speed and health upgrades of the user.
-Multiple flags for checking if the player has I-frames, can move, can attack, can jump.
-Also changed the Entity class's speed definition to a float, since that value will influence a Vector2 value (a float).

+++ src/Entity.cpp
@@ -1,6 +1,6 @@
 #include "Entity.hpp"
 
-Entity::Entity(int hp, int spd, Vector2 pos, const char *sheetLocation)
+Entity::Entity(int hp, float spd, Vector2 pos, const char *sheetLocation)
 {
     health = hp;
     speed = spd;
@@ -19,7 +19,7 @@ int Entity::gethealth()
     return health;
 }
 
-int Entity::getspeed()
+float Entity::getspeed()
 {
     return speed;
 }
