+++ src/Entity.hpp
@@ -6,18 +6,18 @@ class Entity
     protected:
 
         int health;
-        int speed;
+        float speed;
         Vector2 position;
         Texture2D spriteSheet;
         Rectangle hitbox;
     public:
-        Entity(int hp, int spd, Vector2 pos, const char *sheetLocation);
+        Entity(int hp, float spd, Vector2 pos, const char *sheetLocation);
 
         void draw(int frame, int action);
 
         int gethealth();
         void sethealth(int new_health);
-        int getspeed();
+        float getspeed();
         void setspeed(int new_speed);
 
 };
